Delete notes_from_PhillySF_20190426.txt
Student is ranting about how they previously used Arduino successfully, but now they are unable to connect to it and they are not sure if it is related to powering it from a 9V regulator or not. They then indicate that they installed a new version of Anaconda and found that the 'serial' module was not installed by default, and they are suggesting to install 'pyserial' to fix this issue. Finally, they seem to be relieved that not having the position sensor connected is okay.